Isolate queue module changes for schedule_fuzz with proper limits
